+++ compare-frontend/src/app/helpers/d3legend.ts
@@ -1,22 +1,35 @@
-import { color, hsl } from 'd3-color';
+import { Selection } from 'd3-selection';
+import { HSLColor, RGBColor, color, hsl } from 'd3-color';
 import { scaleLinear } from 'd3-scale';
 import { fraction } from './colorhelpers';
+import { rgb } from 'd3-color';
 
 
+function ensureRgb(colorString: string, fallback: {r: number, g: number, b: number} = {r: 125, g: 125, b: 125}): RGBColor {
+  const c = color(colorString);
+  if (c !== null) return c.rgb();
+  return rgb(fallback.r, fallback.g, fallback.b);
+}
+
+export interface Stop {
+  offset: number,
+  color: string,
+  opacity: number
+}
 
 function linearGradientComponent(gradientName: string) {
-  let _stops = [];
+  let _stops: Stop[] = [];
   let _name = gradientName;
-  let _direction = 'horizontal';
+  let _direction: LegendDirection = 'horizontal';
 
-  function _getOrAddDefs(selection) {
-    let defs = selection.select('defs');
+  function _getOrAddDefs(selection: Selection<any, any, any, any>) {
+    let defs: Selection<SVGDefsElement, any, any, any> = selection.select('defs');
     if (!defs.node()) defs = selection.append('defs');
     return defs;
   }
 
-  function _getOrAddGrad(defsSelection) {
-    let grad = defsSelection.select(`#${_name}`);
+  function _getOrAddGrad(defsSelection: Selection<SVGDefsElement, any, any, any>) {
+    let grad: Selection<SVGLinearGradientElement, any, any, any> = defsSelection.select(`#${_name}`);
     if (!grad.node()) {
       grad = defsSelection.append('linearGradient')
         .attr('id', _name);
@@ -34,19 +47,19 @@ function linearGradientComponent(gradientName: string) {
     return grad;
   }
 
-  function gradient(selection) {
+  function gradient(selection: Selection<HTMLElement, any, any, any>) {
     const defs = _getOrAddDefs(selection);
     const gradient = _getOrAddGrad(defs);
     gradient.selectAll('stop').data(_stops)
       .enter()
       .append('stop')
       .attr('offset', d => `${d.offset}%`)
-      .attr('style', d => `stop-color:rgb(${color(d.color).r}, ${color(d.color).g}, ${color(d.color).b}); stop-opacity:${d.opacity}`);
+      .attr('style', d => `stop-color:rgb(${ensureRgb(d.color).r}, ${ensureRgb(d.color).g}, ${ensureRgb(d.color).b}); stop-opacity:${d.opacity}`);
   }
 
-  gradient.stops = (stops) => { _stops = stops; return gradient; }
-  gradient.gradId = (name) => { _name = name; return gradient; }
-  gradient.direction = (direction) => { _direction = direction; return gradient; }
+  gradient.stops = (stops: Stop[])                  => { _stops = stops; return gradient; }
+  gradient.gradId = (name: string)                  => { _name = name; return gradient; }
+  gradient.direction = (direction: LegendDirection) => { _direction = direction; return gradient; }
 
   return gradient;
 }
@@ -54,13 +67,13 @@ function linearGradientComponent(gradientName: string) {
 function scaleColor() {
 
   // default parameters
-  let _domain = [0, 1];
-  let _range = ['rgb(255, 0, 0)', 'rgb(0, 0, 255)'];
+  let _domain: number[] = [0, 1];
+  let _range: string[] = ['rgb(255, 0, 0)', 'rgb(0, 0, 255)'];
 
   // actual interpolation
-  function interpolate(rStart, rEnd, rel) {
-    const cStart = color(rStart);
-    const cEnd = color(rEnd);
+  function interpolate(rStart: string, rEnd: string, rel: number) {
+    const cStart = color(rStart)!;
+    const cEnd = color(rEnd)!;
     const cStartHSL = hsl(cStart);
     const cEndHSL = hsl(cEnd);
     const hInterp = cStartHSL.h + rel * (cEndHSL.h - cStartHSL.h);
@@ -71,7 +84,7 @@ function scaleColor() {
   }
 
   // deciding which colors to interpolate between
-  function scale(d) {
+  function scale(d: number) {
     // if smaller than domain-start ...
     if (d < _domain[0]) {
       return _range[0];
@@ -92,8 +105,8 @@ function scaleColor() {
   };
 
   // public methods
-  scale.domain = function (d) { _domain = d; return scale; };
-  scale.range  = function (r) { _range = r;  return scale; };
+  scale.domain = function (d: number[]) { _domain = d; return scale; };
+  scale.range  = function (r: string[]) { _range = r;  return scale; };
 
   return scale;
 }
@@ -105,9 +118,17 @@ export interface LegendEntryDiscrete {
   text: string
 }
 
+interface LegendEntryDiscretePositioned extends LegendEntryDiscrete {
+  startPos: number,
+  textPos: number,
+  endPos: number,
+  scaledSize: number
+}
+
 
 function createLegendDiscrete(
-  svgSelection, id: string,
+  svgSelection: Selection<HTMLElement, any, any, any>, 
+  id: string,
   width: number, height: number,
   direction: LegendDirection,
   entries: LegendEntryDiscrete[],
@@ -118,7 +139,7 @@ function createLegendDiscrete(
   // layouting
   const dominantSize = direction === 'horizontal' ? width : height;
 
-  const layoutEntries: any[] = [];
+  const layoutEntries: LegendEntryDiscretePositioned[] = [];
   let wSum = 0.0;
   for (const entry of entries) {
     layoutEntries.push({
@@ -166,7 +187,7 @@ function createLegendDiscrete(
       .attr('width',  d => direction === 'horizontal' ? d.scaledSize : graphicEntryWidth)
       .attr('height', d => direction === 'horizontal' ? graphicEntryHeight : d.scaledSize)
       .attr('fill',   d => d.color)
-      .attr('stroke', d => color(d.color).darker());
+      .attr('stroke', d => [... Object.values(ensureRgb(d.color).darker().rgb())] );
 
 
   const labels = legendGroup.selectAll('.label')
@@ -193,7 +214,8 @@ export interface LegendEntryContinuous {
 export type LegendDirection = 'horizontal' | 'vertical';
 
 function createLegendContinuous(
-  svgSelection, id: string,
+  svgSelection: Selection<HTMLElement, any, any, any>,
+  id: string,
   width: number, height: number,
   direction: LegendDirection,
   entries: LegendEntryContinuous[],
@@ -238,7 +260,7 @@ function createLegendContinuous(
     .stops(entries.map(e => ({
       offset: percentageScale(e.position),
       color: colorScale(e.position),
-      opacity: 1
+      opacity: 1,
     })));
 
   svgSelection.call(legendGradient);
@@ -248,10 +270,10 @@ function createLegendContinuous(
     .enter()
     .append('line')
     .attr('class', 'markLine')
-    .attr('x1', d => direction == 'vertical' ? 0                         : placementScale(d.position) )
-    .attr('x2', d => direction == 'vertical' ? graphicWidth + 8           : placementScale(d.position) )
-    .attr('y1', d => direction == 'vertical' ? placementScale(d.position) : 0                          )
-    .attr('y2', d => direction == 'vertical' ? placementScale(d.position) : graphicHeight + 6          )
+    .attr('x1', (d: LegendEntryContinuous) => direction == 'vertical' ? 0                         : placementScale(d.position) )
+    .attr('x2', (d: LegendEntryContinuous) => direction == 'vertical' ? graphicWidth + 8           : placementScale(d.position) )
+    .attr('y1', (d: LegendEntryContinuous) => direction == 'vertical' ? placementScale(d.position) : 0                          )
+    .attr('y2', (d: LegendEntryContinuous) => direction == 'vertical' ? placementScale(d.position) : graphicHeight + 6          )
     .attr('stroke', 'grey');
 
   const graphic = legendGroup.append('rect')
@@ -265,11 +287,11 @@ function createLegendContinuous(
     .enter()
     .append('g')
       .attr('class', 'label')
-      .attr('transform', d => direction === 'horizontal' ?
+      .attr('transform', (d: LegendEntryContinuous) => direction === 'horizontal' ?
         `translate(${placementScale(d.position)}, ${graphicHeight + 15})` :
         `translate(${graphicWidth + 10}, ${placementScale(d.position)})`)
     .append('text')
-      .text(d => d.text)
+      .text((d: LegendEntryContinuous) => d.text)
       .style('text-anchor', direction === 'horizontal' ? 'middle' : 'left')
       .style('dominant-baseline', 'middle');
 
@@ -326,7 +348,7 @@ export function legendComponent() {
   let _margin = 10;
   let _entries: LegendEntry[] = [];
 
-  function legend(selection) {
+  function legend(selection: Selection<HTMLElement, any, any, any>) {
     if (_continuous) {
       
       let p = 0;